refactor: remove quox→jsx dependency
