+++ packages/jsx/jsx-runtime_test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "@std/assert";
-import { createVNode, Fragment, isQuoxVNode, type QuoxComponent, serializeQuoxStyle } from "./jsx-runtime.ts";
+import { createVNode, Fragment, isQuoxVNode, type QuoxComponent } from "./jsx-runtime.ts";
 
 Deno.test("createVNode stores type, props, children, and key", () => {
   const vnode = createVNode("h1", { id: "title", children: "Hello" }, "main");
@@ -16,19 +16,6 @@ Deno.test("Fragment is available as a VNode type", () => {
 
   assertEquals(vnode.type, Fragment);
   assertEquals(vnode.children, ["a", null, false, "b"]);
-});
-
-Deno.test("serializeQuoxStyle preserves custom properties and kebab-cases normal properties", () => {
-  assertEquals(
-    serializeQuoxStyle({
-      "--accent": "red",
-      backgroundColor: "var(--accent)",
-      width: 100,
-      WebkitLineClamp: 2,
-      display: null,
-    }),
-    "--accent:red;background-color:var(--accent);width:100;-webkit-line-clamp:2",
-  );
 });
 
 Deno.test("QuoxComponent accepts an async implementation", () => {
